#176: Extract fp Stream to separate script
Also:
* Add $api.Function.RegExp.exec
* Remove redundant fifty.$api.Function
* Add commented-out code beginning #176
