+++ loader/$api.js
@@ -72,7 +72,7 @@
 
 		(function() {
 			var old = code.Function_old({ deprecate: $exports.deprecate });
-			var current = code.Function({ $api: $exports, events: events, old: old, deprecate: $exports.deprecate });
+			var current = code.Function({ $api: $exports, events: events, old: old, deprecate: $exports.deprecate, script: script });
 			Object.assign($exports, current);
 		})();
